project-root/requirements.txt
dj_database_url==2.2.0
Django==5.1
djangorestframework==3.15.2
gunicorn>=20.0.4
django-cors-headers>=4.4.0
psycopg>= 3.2.1
psycopg2>=2.9.9